+++ c3smembership/new_member.py
@@ -53,12 +53,12 @@ def new_member(request):
         )
         lastname = colander.SchemaNode(
             colander.String(),
-            title=u'Nachnahme (b. Körpersch.: c/o Person)',
+            title=u'Nachname (b. Körpersch.: c/o Person)',
             oid="lastname",
         )
         email = colander.SchemaNode(
             colander.String(),
-            title=_(u'Email'),
+            title=_(u'E-Mail'),
             validator=colander.Email(),
             oid="email",
         )
@@ -70,12 +70,12 @@ def new_member(request):
         )
         address1 = colander.SchemaNode(
             colander.String(),
-            title='Addresse Zeile 1'
+            title='Adresse Zeile 1'
         )
         address2 = colander.SchemaNode(
             colander.String(),
             missing=unicode(''),
-            title='Addresse Zeile 2'
+            title='Adresse Zeile 2'
         )
         postcode = colander.SchemaNode(
             colander.String(),